refactor: remove HealthCheckController
