+++ src/main/resources/templates/health.html
@@ -1,0 +1,1 @@
+ok
